requirements/requirement_open_source.txt
torch>=2.7.0
torchvision>=0.22.0
diffusers @ git+https://github.com/huggingface/diffusers.git
transformers>=4.55.0
accelerate>=1.9.0
safetensors>=0.7.0
sentencepiece>=0.2.0
einops>=0.8.0
timm>=1.0.0
peft>=0.17.0
optimum-quanto>=0.2.7
Pillow>=12.0.0
opencv-python>=4.12.0
scikit-image>=0.25.0
huggingface-hub>=0.34.0
datasets>=4.0.0
numpy>=2.0.0
tqdm>=4.67.0
requests>=2.32.0
PyYAML>=6.0.0
qwen-vl-utils>=0.0.11